Introduces new field stationType on node configurations, remove old files and workflows
This commit introduces new field stationType on node configurations for stations, removes several outdated and unnecessary files and workflows from the codebase. Specifically, it deletes the Changelog, the check-generated and cometbft-docker Github workflows, the Code of Conduct, and the Contributing guide. By removing these files, the project is kept tidy and easier to navigate. Any relevant historical information is still available in the repository history.